Moved WatsonOverviewWidget to widgets mod
